CIL-1413 Better error logging for "Failed to getuser"
In handleGotUser(), there is a sanity check to make sure that there was
no dbService error and that the user_uid is in the session. (This check
also handles if eduGAIN IdP tried to get X509 cert, but that is already
logged.) A simple "Failed to getuser" error message is logged, but that
is now improved by also logging the dbService status code and the
user_uid.