CIL-1369 Move SSO code to a Util function.
All of the code for checking SSO that was in Content.php has been moved
to Util.php. This is because the code needs to be called twice:
(1) Before the IdP list is shown, detect if this transaction is eligible
for SSO, and if so check if a previously used IdP matches the current
session IdP.
(2) After the user has logged in to an IdP (for the first time in the
session), save that IdP to the session for future SSO checks.