CIL-2188 Don't show warnings for missing name/email
When the user's IdP doesn't assert name or email address, we show a
warning message that the user won't be able to generate a certificate.
We also show little warning icons for the missing attributes. This is
confusing to some users who just need to use CILogon OIDC capabilities,
but still go to https://cilogon.org to view their user attributes.
So change the warning message and don't show the warning icons for
missing name and/or email address.